ai-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
groq==0.4.2
chromadb==0.3.21
requests==2.31.0
gunicorn==21.2.0
sentence-transformers==2.2.2
langchain==0.1.16
pypdf==4.0.1
python-docx==0.8.11
chardet==5.2.0
pytest==7.4.3
pytest-mock==3.12.0
